Moved duplicated classes to tweet-stepengine project. References to TweetStreamDataProvider and the various forms of Step triggering TweetStreamDataProvider have been modified to use those implementations in project tweet-stepengine fixes TweetWallFX/TweetwallFX#404
