+++ devoxx-2017-be/cinema/src/main/resources/META-INF/services/org.tweetwallfx.controls.dataprovider.DataProvider$Factory
@@ -1,3 +1,2 @@
 org.tweetwallfx.devoxx2017be.dataprovider.TopTalksTodayDataProvider$Factory
 org.tweetwallfx.devoxx2017be.dataprovider.TopTalksWeekDataProvider$Factory
-org.tweetwallfx.devoxx2017be.dataprovider.TweetStreamDataProvider$Factory
