+++ devoxx-2018-pl/cinema/src/main/resources/META-INF/services/org.tweetwallfx.controls.dataprovider.DataProvider$Factory
@@ -1,3 +1,2 @@
 org.tweetwallfx.devoxx18pl.dataprovider.TopTalksTodayDataProvider$Factory
 org.tweetwallfx.devoxx18pl.dataprovider.TopTalksWeekDataProvider$Factory
-org.tweetwallfx.devoxx18pl.dataprovider.TweetStreamDataProvider$Factory
