--- voxxed-days-zurich/src/main/resources/META-INF/services/org.tweetwallfx.controls.dataprovider.DataProvider$Factory
@@ -1,1 +1,0 @@
-org.tweetwallfx.vdz.dataprovider.TweetStreamDataProvider$Factory
